Rename udf files
